+++ src/api/notification.js
@@ -7,6 +7,8 @@ import {
 
 const { API_KEY } = API.PROPOSAL;
 const { NOTIFICATION_API_URL } = API.NOTIFICATION;
+const ENDPOINT_MARK_READ = 'read';
+const ENDPOINT_MARK_ALL_READ = 'mark-all-read';
 
 const getAxiosConfig = () => ({
   headers: {
@@ -15,68 +17,38 @@ const getAxiosConfig = () => ({
   }
 });
 
-const tempDummyData = [
-  {
-    id: '01',
-    preference_id: '03',
-    user_email: 'ashiq_sultan@iqvia.com',
-    created_at: '2022-07-01',
-    data: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
-    isSeen: false,
-    proposal_id: '32bc1cf9-66ed-4c10-9c72-f1195c4b1668',
-    opportunity_no: 'UZA85791',
-    action_url: 'http://localhost:8080/opportunities/UZA85791'
-  },
-  {
-    id: '02',
-    preference_id: '04',
-    user_email: 'ashiq_sultan@iqvia.com',
-    created_at: '2022-07-01',
-    data: 'foo bar 02',
-    isSeen: true,
-    proposal_id: '32bc1cf9-66ed-4c10-9c72-f1195c4b1668',
-    opportunity_no: 'UZA85791',
-    action_url: 'http://localhost:8080/opportunities/UZA85791'
-  },
-  {
-    id: '03',
-    preference_id: '05',
-    user_email: 'ashiq_sultan@iqvia.com',
-    created_at: '2022-07-03',
-    data: 'lorem ipsum 03',
-    isSeen: false,
-    proposal_id: '32bc1cf9-66ed-4c10-9c72-f1195c4b1668',
-    opportunity_no: 'UZA85791',
-    action_url: 'http://localhost:8080/opportunities/UZA85791'
-  }
-];
-
 function fetchNotifications() {
   const email = getUserEmail();
   const config = getAxiosConfig();
   return new Promise((resolve, reject) => {
-    // TODO
-    // Make API call to get all notifications
-    resolve(tempDummyData);
+    axios
+      .get(`${NOTIFICATION_API_URL}`, config)
+      .then(response => resolve(response.data?.data))
+      .catch(err => reject(err));
   });
 }
 
 function updateSeenOne(notificationId) {
   const config = getAxiosConfig();
   return new Promise((resolve, reject) => {
-    // TODO
-    // Make API call to update one notification as seen
-    resolve([]);
+    axios
+      .patch(
+        `${NOTIFICATION_API_URL}/${notificationId}/${ENDPOINT_MARK_READ}`,
+        {},
+        config
+      )
+      .then(response => resolve(response.data?.data))
+      .catch(err => reject(err));
   });
 }
 
-function updateSeenBatch(notificationIds: Array<string>) {
+function updateSeenBatch() {
   const config = getAxiosConfig();
-  const data = { notificationIds };
   return new Promise((resolve, reject) => {
-    // TODO
-    // Add api call to update all notifiaction of an user as seen
-    resolve([]);
+    axios
+      .patch(`${NOTIFICATION_API_URL}/${ENDPOINT_MARK_ALL_READ}`, {}, config)
+      .then(response => resolve(response.data?.data))
+      .catch(err => reject(err));
   });
 }